examples/workers/tutorial_pull_tweeter_hashtags/requirements.txt
requests
python-dateutil
python-twitter
